% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookiecutter.r
\name{bake}
\alias{bake}
\title{Fill a directory with files based on a cookiecutter-based template.}
\usage{
bake(template, output_dir, extra_context = list(), overwrite = F, no_input = F)
}
\arguments{
\item{template}{Template to apply}

\item{output_dir}{Directory to fill}

\item{extra_context}{Context variables to use while populating template}

\item{overwrite}{Whether to overwrite existing files}

\item{no_input}{Don't prompt for context variables
specified in cookiecutter.json}
}
\description{
This function takes a cookiecutter-compatible template directory and
resolves template placeholders in both filenames and (for text files)
content.
}
\author{
Felix Henninger
}
