% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookiecutteR-package.R
\docType{package}
\name{cookiecutteR-package}
\alias{cookiecutteR}
\alias{cookiecutteR-package}
\title{cookiecutteR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sean Warlick \email{warlick.sean@gmail.com}

}
\keyword{internal}
