% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_create_utils.R
\name{copy_template}
\alias{copy_template}
\title{Use A Template to Add Project Infastructure.}
\usage{
copy_template(path, template, data = list())
}
\arguments{
\item{path}{path where file should be added.}

\item{template}{Name of the template to use.  Uses the templates included in \code{\link[usethis]{use_usethis}}.}

\item{data}{a list providing data, like author or project name to add to
the template.}
}
\description{
For internal use. Users should call \code{\link[usethis]{use_template}}
instead.
}
\keyword{internal}
