% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_create_utils.R
\name{create_file_structure}
\alias{create_file_structure}
\title{Create Project Directories.}
\usage{
create_file_structure(project_root)
}
\arguments{
\item{project_root}{root directory for the project being created.}
}
\description{
Create Project Directories.
}
\keyword{internal}
