% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_create_utils.R
\name{create_makefile}
\alias{create_makefile}
\title{Add a Makefile to the project}
\usage{
create_makefile(project_root)
}
\arguments{
\item{project_root}{root directory for the project being created.}
}
\description{
For internal use. Users should rely on funtions
\code{\link[usethis]{use_travis}} rather than calling \code{create_makefile}
directly.
}
\keyword{internal}
