% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_project.R
\name{create_new_project}
\alias{create_new_project}
\title{Create A New Project}
\usage{
create_new_project(path)
}
\arguments{
\item{path}{directory where the project should be created. Don't
include the project name.}
}
\description{
Creates a new project using the \href{https://drivendata.github.io/cookiecutter-data-science/}{Cooiecutter for Data Science} structure.
}
\examples{
\dontrun{
create_new_project()
}
}
