% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_create_utils.R
\name{create_readme}
\alias{create_readme}
\title{Create README.Rmd for Project.}
\usage{
create_readme(project_root)
}
\arguments{
\item{project_root}{root for the project being created.}
}
\value{
Creates a README.Rmd file in the project directory.
}
\description{
For internal use. Users should rely on functions
\code{\link[usethis]{use_readme_rmd}} rather than calling \code{create_readme}
directly.
}
\keyword{internal}
