% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enable_git.R
\name{enable_git}
\alias{enable_git}
\title{Enable a git repository}
\usage{
enable_git(project_root, choice, remote)
}
\arguments{
\item{project_root}{root directory for the project being created}

\item{choice}{the choice selected interactively when
\code{\link{create_new_project}} is called}

\item{remote}{the url of the remote orign.}
}
\description{
An internal method for initalizing a git repository and adding a
remote origin. For internal use. Users should call
\code{\link[git2r]{init}} directly instead.
}
\keyword{internal}
