% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enable_renv.R
\name{enable_renv}
\alias{enable_renv}
\title{Create project specific library using renv.}
\usage{
enable_renv(project_root, renv_status)
}
\arguments{
\item{project_root}{root directory for the project being created.}

\item{renv_status}{a boolean indicating if renv should be utilized. Selected
interactively when \code{\link{create_new_project}} is called.}
}
\description{
Intend for internal use only.  If you want to create a proejct library
outside of the project set up call \code{\link[renv]{init}} directly. User
consents to the creation of renv when selecting the option to enable it.
}
\keyword{internal}
