% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_files.r
\name{generate_files}
\alias{generate_files}
\title{Generate a new directory from a template, given a predefined context}
\usage{
generate_files(
  template_dir,
  output_dir,
  context = list(),
  context_prefix = NA,
  exclude = c(),
  overwrite = FALSE
)
}
\arguments{
\item{template_dir}{Template directory to copy}

\item{output_dir}{Target directory to copy to (defaults to project directory)}

\item{context}{Data available in templates}

\item{context_prefix}{Prefix key to apply to context data (see description)}

\item{exclude}{Files to exclude, as an array of paths with respect to
the template directory}

\item{overwrite}{Whether to overwrite existing files}
}
\description{
Generate a new directory from a template, given a predefined context
}
\keyword{internal}
