% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_project.R
\name{create_file_structure}
\alias{create_file_structure}
\title{Title}
\usage{
create_file_structure(project_name, path)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
