% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_create_utils.R
\name{create_file_structure}
\alias{create_file_structure}
\title{Create Project Directories.}
\usage{
create_file_structure(project_root, check_project_name)
}
\arguments{
\item{project_root}{root directory for the project being created.}

\item{check_project_name}{check if project name is a valid name for CRAN.
Set via config options.}
}
\description{
Create Project Directories.
}
\keyword{internal}
